+++ src/Ookii.CommandLine/NameTransformExtensions.cs
@@ -14,7 +14,8 @@ namespace Ookii.CommandLine
         /// <param name="transform">The transformation to apply.</param>
         /// <param name="name">The name to transform.</param>
         /// <param name="suffixToStrip">
-        ///   An optional suffix to remove from the string before transformation.
+        ///   An optional suffix to remove from the string before transformation. Only used if
+        ///   <paramref name="suffixToStrip"/> is not <see cref="NameTransform.None"/>.
         /// </param>
         /// <returns>The transformed name.</returns>
         /// <exception cref="ArgumentNullException">